Tenant Create and get all tenants is working
Okay so basically, to create a tenant you have to use the user controller. Which means the frontend might be sticky, due to the fact you're going to have to get the Name and Surname information using the UserID field, but if you need some help with that let me know and I'll try figure it out. Going to do the contractor create and employee create next so that all users are working and then moving onto front end property reports.